+++ src/main/java/learn/thyme/thyteams/infrastructure/web/WebMvcConfiguration.java
@@ -1,7 +1,9 @@
 package learn.thyme.thyteams.infrastructure.web;
 
+import learn.thyme.thyteams.user.PhoneNumberFormatter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.format.FormatterRegistry;
 import org.springframework.web.servlet.LocaleResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
@@ -25,5 +27,10 @@ public class WebMvcConfiguration implements WebMvcConfigurer {
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
         registry.addInterceptor(localeInterceptor());
+    }
+
+    @Override
+    public void addFormatters(FormatterRegistry registry) {
+        registry.addFormatter(new PhoneNumberFormatter());
     }
 }